Move SyntaxMap tests to their own file
